1. string_nconcat

+++ 0x0C-more_malloc_free/1-string_nconcat.c
@@ -26,7 +26,7 @@ char *string_nconcat(char *s1, char *s2, unsigned int n)
 	{
 		j++;
 	}
-	if (n < j)
+	if ((int) n < j)
 		j = n; 
 	p = malloc(sizeof(char) * (i + j - 1));
 	if (!p)
